+++ pom.xml
@@ -23,7 +23,7 @@
         <aws-java-sdk-s3.version>1.11.448</aws-java-sdk-s3.version>
         <felles.version>1.0_20191106113227_c2ad869</felles.version>
         <mockk.version>1.9.3</mockk.version>
-        <token-validation-spring.version>1.1.2</token-validation-spring.version>
+        <token-validation-spring.version>1.1.3</token-validation-spring.version>
      </properties>
 
     <dependencies>
@@ -83,14 +83,8 @@
         </dependency>
         <dependency>
             <groupId>no.nav.security</groupId>
-            <artifactId>token-client-spring</artifactId>
+            <artifactId>token-validation-spring</artifactId>
             <version>${token-validation-spring.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>no.nav.security</groupId>
-            <artifactId>token-validation-test-support</artifactId>
-            <version>${token-validation-spring.version}</version>
-            <scope>test</scope>
         </dependency>
 
         <!-- Storage -->
@@ -155,7 +149,13 @@
             <version>1.12.1</version>
             <scope>test</scope>
         </dependency>
-     </dependencies>
+        <dependency>
+            <groupId>no.nav.security</groupId>
+            <artifactId>token-validation-test-support</artifactId>
+            <version>${token-validation-spring.version}</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 
     <build>
         <sourceDirectory>${project.basedir}/src/main/kotlin</sourceDirectory>
