+++ src/main/kotlin/no/nav/familie/dokument/Launcher.kt
@@ -5,7 +5,7 @@ import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
 
 @SpringBootApplication(scanBasePackages = ["no.nav.familie.dokument"])
-@EnableJwtTokenValidation
+@EnableJwtTokenValidation(ignore = ["org.springframework", "springfox.documentation.swagger.web.ApiResourceController"])
 class Launcher
 
 fun main(args: Array<String>) {
