+++ src/main/resources/application-dev.yaml
@@ -1,8 +1,8 @@
 no.nav.security.jwt:
   issuer.selvbetjening:
-    discoveryurl: http://localhost:8080/metadata
-    accepted_audience: aud-localhost
-    cookie_name: localhost-idtoken
+    discoveryurl: http://metadata
+    acceptedAudience: aud-localhost
+    cookiename: localhost-idtoken
 
 FAMILIE_DOKUMENT_STORAGE_ENCRYPTION_PASSWORD: not-a-real-pw
 SOKNAD_PDF_SVG_SUPPORT_GENERATOR_URL: #mockes ut lokalt
